Create Get_Bivalve_OGs.py
To ID and extract the names of orthogroups which have gene duplications retained in at least 50% bivalves analysed. 
When given the Orthofinder output files "Duplications.tsv" and "Orthogroups.GeneCount.tsv", and specifying a taxonomical Node (clade specific tree node), this script produces a tab seperated .csv with the average FoldChange in gene duplication for Bivalves vs Non-Bivalve Molluscs in this dataset.