Create kofamscan_analyses.py
This takes a directory of Kofamscan output files and delivers a tab separated .txt file of the hits sorted by count with avg. E-score and cumulative Score. 
It deals with hits which have passed the score threshold separately (although these are often much rarer and difficult to draw conclusions from).