Create OGfa2graph.py
Takes an Orthogroup.fasta file and runs Blastp all versus all to generate a pairwise .txt file.
This can be used to generate a graph using Graphia. Nodes are also annotated according to species and class. 
Requires Blastp module to be loaded, and the ability to create a megablastDB (callable from cmd line)